Graph-Enhanced Agentic RAG for Corporate Intelligence/requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
neo4j>=5.15.0

# Document Processing
PyMuPDF>=1.23.0
pdfminer.six>=20221105
pypdf2>=3.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization
plotly>=5.17.0
networkx>=3.2.1
pyvis>=0.3.2
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
tqdm>=4.66.0

# Optional: For Gemma model (if using local deployment)
# accelerate>=0.25.0
# bitsandbytes>=0.41.0
